welding_server/backend/requirements.txt
# Django Core
Django
djangorestframework
django-cors-headers

# Database
# psycopg2-binary  # Only needed for PostgreSQL (production)

# Image Processing
Pillow

# PDF Generation
reportlab

# PDF Parsing (for attendance sheet import)
pdfplumber

# SSH/SCP for MLOps Deployment
paramiko

# System Monitoring
psutil

# Utilities
python-dotenv
